Add Data Officers Add Authorization Control Definitions Changes in UI

+++ ckanext/definitions/logic/action/delete.py
@@ -3,6 +3,7 @@
 '''API functions for deleting data from CKAN.'''
 
 import logging
+from ckan.plugins import toolkit
 
 import sqlalchemy as sqla
 
@@ -26,7 +27,6 @@ validate = ckan.lib.navl.dictization_functions.validate
 # actions in the action API.
 ValidationError = ckan.logic.ValidationError
 NotFound = ckan.logic.NotFound
-_check_access = ckan.logic.check_access
 _get_or_bust = ckan.logic.get_or_bust
 _get_action = ckan.logic.get_action
 
@@ -54,7 +54,27 @@ def definition_delete(context, data_dict):
     if definition_obj is None:
         raise NotFound(_('Could not find definition "%s"') % definition_id)
 
-    # _check_access('definition_delete', context, data_dict)
+    toolkit.check_access('definition_delete', context, data_dict)
 
     definition_obj.delete()
     model.repo.commit()
+
+
+def data_officer_delete(context, data_dict):
+    '''
+    Removes the role Data Officer from a User
+    :param context:
+    :param data_dict: contains 'user_id'
+    :return: the definition added to the DB
+    '''
+
+    user_id = data_dict['user_id']
+    user_extras = toolkit.get_action('user_extra_show')(context, {"user_id": user_id})['extras']
+
+    for extra_dict in user_extras:
+        if extra_dict['key'] == 'Data Officer':
+            _data_dict = {"user_id": user_id, "extras": [{"key":"Data Officer", "new_value":""}]}
+            result = toolkit.get_action('user_extra_update')(context, _data_dict)
+            return "User removed Successfuly from the Data Officers List."
+    return "User is not a Data Officer"
+
